Create revit_to_robot_interface_dynamo.py
Initial version commit. Basic outline of project.

common language references are added along with imports in boilerplate as required.
add_handler_to_events function is called by main, providing the handler function as well as the DocumentSaved and DocumentSavedAs events. This function adds the function to the list of functions called when these events occur.

handler function intends to export the active document to robot structural analysis, however as this is a first draft, it instead uses a simpler command (Print Preview) for testing purposes.